[changed]changed type of variables

+++ src/binarization.rs
@@ -28,7 +28,7 @@ fn conv_from_line(img: &Vec<u8>, x_max: usize) -> Vec<Vec<u8>> {
 }
 
 // threshold: 輝度のしきい値（これ以上は白画素，これ以下は黒画素）
-fn binarize(img: &mut Vec<Vec<u8>>, threshold: u8) {
+fn binarize(img: &mut Vec<Vec<u8>>, threshold: usize) {
     let mut p_src = conv_to_line(img);
     let x_max = img[0].len();
     let y_max = img.len();
@@ -39,6 +39,6 @@ fn binarize(img: &mut Vec<Vec<u8>>, threshold: u8) {
         lut[i] = 255;
     }
     for i in 0..x_max * y_max {
-        img[i / x_max][i % x_max] = lut[p_src[i]];
+        img[i / x_max][i % x_max] = lut[p_src[i] as usize];
     }
 }
